requirements.txt
streamlit==1.48.1
yfinance==0.2.36
pandas==2.3.1
plotly==6.3.0
python-dotenv==1.0.0
requests==2.32.4
numpy==2.3.2
ta==0.10.2 